Ability to collect apples then add to inventory.
When near an apple tree when you press the 'E' key, you will be able to collect an apple. The growth rate of the apple trees is 3 seconds, but we can tweak that later on.

Press the 'i' key to see the inventory update with the apple(s).

I intend to use these apples for healing the player.

We can probably also add more food options that gives a wide variety of effectiveness for health and add other items in the inventory like weapons that each adds a variety of damage.