Created an NPC
I created an NPC that walks around the map.

The NPC art asset itself is a placeholder until we can find better fitting character sprites for the game.

I'll make a few tweaks for it to stay in its specific region.

However, although I've added a collision shape to the character it is able to walk pass through trees and buildings. Fortunately it's able to collide with the player.

Other regions have this issue as well for the player too.

Maybe Zachary can fix this issue.

Furthermore, pressing the 'c' key will allow the player to interact with the NPC.

As of now, it should stop moving and stand by the player when pressing the key.

I'll continue working on the dialogue system next, along with updating the player hp and use of inventory items.