Added Village Chief NPC / Began tutorial sequence
I've reworked the main scene a little to begin a tutorial of the game and added a Village Chief NPC to introduce turn-based combat.

I created a textbox that follows the player on which provides how the controls work.

The issue is that the textbox reappears all over again after changing a scene. For example, after a turn-based combat with a slime, bat, or spider.

I should be able to resolve this issue tomorrow and continue working on the tutorial/story.

I'll also continue to implement player hp and recover it with the use of food items in the in the inventory.

I'll also add food items specific to each region that adds a variety of hp points laid around their specific region for the player to search and collect. Either this week or next week.

I've also added music to the main menu and a fade transition between the main menu and main scene.