Player HP get regenerate using inventory items
The Player HP should remain consistent between the main.tscn and combat.tscn.

I added a Player HP label below the player at all times to see HP.

The Player is now able to recover its HP by selecting an item in the Inventory.

As of now,
The apple item increases HP by +5
The health potion item increases HP by +10

I also fixed the opening textbox, so it should only appear the first time.

I also fixed the dialogue boxes with the NPCs. Before pressing the 'C' key wherever you were would activate the dialogue with all NPCs in the same time.

Now the Player has to be near the NPC to talk with them individually.

Next, I'm going to have the Enemy Hp be consistent, so after a battle, then resuming it, it doesn't reset their HP back to full capacity.

After that, I'm going to work on 3 different enemies for the corruption region and add another item.