Finished tutorial and added NPCs for Desert and Corruption Region
I think I finalized all the dialogue with the NPCs, you can check the .json files if you think there can be improvements.

I added an NPC for the Desert Region to further the plot and introduce the health potion, which will later be add around the Desert region for the player to collect.

I also added a survivor NPC for the Corruption Region to further the plot and introduce the herb item, which will later be add around the Corruption region for the player to collect.